+++ test/RunLengthTextureFeaturesImageFilterTestSeparateFeatures.cxx
@@ -47,7 +47,7 @@ int RunLengthTextureFeaturesImageFilterTestSeparateFeatures( int argc, char *arg
   const unsigned int VectorComponentDimension = 10;
 
   // Declare types
-  typedef int                                         InputPixelType;
+  typedef float                                       InputPixelType;
   typedef float                                       OutputPixelComponentType;
   typedef itk::Vector< OutputPixelComponentType, VectorComponentDimension >
                                                       OutputPixelType;
